backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.0
redis==5.0.1
rq==1.16.0
gunicorn==21.2.0
email-validator==2.1.0
cryptography==42.0.0
openpyxl==3.1.2
reportlab>=4.0.0
qrcode[pil]>=7.0.0
pillow>=10.0.0
psycopg2-binary>=2.9.9
psycopg2-binary>=2.9.9
